pkg/store/metadatastore: add DatabaseStore
In theory, this should be more reliable when it comes to deploying
multiple instances of nix-casync.

In practice, it's a bit hardcoded to sqlite (right now), and I didn't
(yet) win the fight against bun to get all the relationships described.

Let's put this into the open, maybe someone with more experience with
Bun knows how to fix this ;-)