NarStore: name objects after their content hash/narhash
This changes the interface of the NarinfoStore to not accept a "narhash
to store it at" upfront, but instead lets the stores provide the hash of
the file (after everything has been written).

This will make NarStore a fully content-addressed store.

We want to verify hashes for Narfiles while uploading anyways.