pkg/store/casync_store_{reader,writer}.go: remove buggy init() functions
This broke the Close() function if it was called before the first
Read(…) call.

Move basic initialization into new NewCasyncStoreNarReader,
NewCasyncStoreNarWriter functions, and in the case of
casyncStoreNarReader, do the assemble on the first Read() call.

This also has the advantage of simplifying the error handling code in
casyncStoreNarReader, as we can simply return the error, and leave the
tmpFile removal up to the caller, who needs to run Close().