pkg/server/server.go: Don't update NarMeta when receiving the same .nar file again
Right now, we don't do any reference scanning on the uploaded .nar files
ourselves, but initialize it with an empty list of references, and fill
it whenever we receive a .narinfo file referring to that narfile, if
there's (still) an empty list.

A Nix client might sometimes upload the same .nar file twice (which can
happen more easily when uploads happen with compression, as we always
reply with a 404 if Nix asks for $filehash.nar.$compression (we unpack
on upload and don't keep track of the hashes of compressed files)

Whenever this happens, we previously removed the References list. If
then the client cancels (and doesn't get to upload the .narinfo), or
in general while it didn't upload the .narinfo again, nix-casync would
reply with an empty list of references, which is not what we want.