pkg/server/server.go: fix uploading self-references
Previously, nix-casync didn't properly handle uploaded store paths which
contained self-references.

Initially, this silently emptied the references list, which was fixed in
05911002ac3cdc39f5de3e060842636baa35b656. With this commit, nix-casync
rejected the upload.

This commit now includes the fix - if we receive a .narinfo with
references (and the NarMeta doesn't contain reference data yet), we
write PathInfo first, then update NarMeta with the received references
list.

This means, if someone immediately queries the uploaded .narinfo while
the NarMeta hasn't been updated, it would not contain the references
yet.

This should be migitated by adding a transactional interface to
metadatastore, which is not implemented.

Fixes #30.